requirements.txt
pandas==2.3.3
streamlit==1.52.1
plotly-express==0.4.1